stage1-module4-collections-task3-map/src/main/java/com/epam/mjc/collections/map/FunctionValueFinder.java
package com.epam.mjc.collections.map;

import java.util.HashMap;
import java.util.List;
import java.util.Map;

public class FunctionValueFinder {
    public boolean isFunctionValuePresent(List<Integer> sourceList, int requiredValue) {
        Map<Integer, Integer> functionMap = calculateFunctionMap(sourceList);

        return functionMap.containsValue(requiredValue);
    }

    private Map<Integer, Integer> calculateFunctionMap(List<Integer> sourceList) {
        Map<Integer, Integer> newMap = new HashMap<>();
        for (Integer i : sourceList) {
            newMap.put(i, 5 * i + 2);
        }
        return newMap;
    }
}
